+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Angry creepers now blow up after death like 0.30 creepers
+  * Configurable to also make charged creepers only or any creeper to behave the same
 * Pearler and Fisher Mobs can now use the item as long as they are not underwater (they couldn't use them if touching water)
 * Pearler mobs no longer use the pearl if they can't see the target
 * Pearler mobs cooldown to throw a pearl increased